// lib/user/api/usertypes.ts
export type BubbleUser = {
  uniqueID: string;
  firstName: string;
  lastName: string;
  timezone: string;
  email: string;
  verifiedPhone: string;
  bubbleRole: string;
  stripeCustomerId: string;
  stripeSubscriptionId: string;
  stripePriceId: string;
  stripeCurrentPeriodEnd: string;
  accountabilitySetting: any;
  challenges: any;
  currentRoutine: any;
  currentNutrition: any;
  fitnessGoals: any;
  height: number;
  weight: number;
  age: number;
  preferredExercises: any;
  targetRoutine: any;
  wakeupTime: any;
  sleepTime: any;
  userReady: any;
};
export type BubbleUserProfile = {
  accountabilitySetting: any;
  challenges: any;
  currentRoutine: any;
  currentNutrition: any;
  fitnessGoals: any;
  height: number;
  weight: number;
  age: number;
  preferredExercises: any;
  targetRoutine: any;
  wakeupTime: any;
  sleepTime: any;
  userReady: any;
};